utils/requirements.txt
pyquery
bibtexparser
selenium
nltk
pyyaml
